chore(notification): Move the Notification directory structure
