+++ packages/element3/src/components/Notification/index.js
@@ -1,4 +1,4 @@
-import { Notification } from './Notification.js'
+import { Notification } from './src/Notification.js'
 
 export function useNotify() {
   return Notification
