+++ packages/element3/src/components/Notification/tests/Notification.spec.js
@@ -1,4 +1,4 @@
-import { Notification } from '../Notification.js'
+import { Notification } from '../src/Notification.js'
 import { h } from 'vue'
 
 describe('Notification.js', () => {
